+++ cruft/_cli.py
@@ -1,7 +1,7 @@
 """This module defines CLI interactions when using `cruft`."""
 import json
 from pathlib import Path
-from typing import Optional
+from typing import List, Optional
 
 import typer
 
@@ -100,6 +100,9 @@ def create(
         show_default=False,
         help="Overwrite the contents of the output directory if it already exists",
     ),
+    skip: Optional[List[str]] = typer.Option(
+        None, "--skip", show_default=False, help="Default files/pattern to skip on update"
+    ),
 ) -> None:
     _commands.create(
         template_git_url,
@@ -111,6 +114,7 @@ def create(
         directory=directory,
         checkout=checkout,
         overwrite_if_exists=overwrite_if_exists,
+        skip=skip,
     )
 
 
@@ -225,8 +229,8 @@ def update(
         False,
         "--allow-untracked-files",
         help=(
-                "Allow the project's cruft to be updated if there are untracked files in the git"
-                " repository (but no other changes)"
+            "Allow the project's cruft to be updated if there are untracked files in the git"
+            " repository (but no other changes)"
         ),
     ),
 ) -> None: