+++ tests/test_cli.py
@@ -157,6 +157,41 @@ def test_update_unclean(cruft_runner, cookiecutter_dir):
     assert result.exit_code == 1
 
 
+def test_update_allow_untracked_files(cruft_runner, cookiecutter_dir):
+    run(["git", "init"], cwd=cookiecutter_dir)
+    run(["git", "add", "-A"], cwd=cookiecutter_dir)
+    run(
+        [
+            "git",
+            "-c",
+            "user.name='test'",
+            "-c",
+            "user.email='user@test.com'",
+            "commit",
+            "-am",
+            "test",
+        ],
+        cwd=cookiecutter_dir,
+    )
+    run(["touch", "new_file.txt"], cwd=cookiecutter_dir)
+    result = cruft_runner(["update", "--project-dir", str(cookiecutter_dir), "-y"])
+    assert "Cruft cannot apply updates on an unclean git project." in result.stdout
+    assert result.exit_code == 1
+    result = cruft_runner(
+        [
+            "update",
+            "--project-dir",
+            str(cookiecutter_dir),
+            "-y",
+            "--allow-untracked-files",
+            "-c",
+            "updated",
+        ]
+    )
+    assert "cruft has been updated" in result.stdout
+    assert result.exit_code == 0
+
+
 def test_update(cruft_runner, cookiecutter_dir):
     result = cruft_runner(
         ["update", "--project-dir", cookiecutter_dir.as_posix(), "-y", "-c", "updated"]
@@ -203,7 +238,7 @@ def test_update_with_conflicts_with_git(cruft_runner, cookiecutter_dir):
     assert result.exit_code == 0
     assert set(cookiecutter_dir.glob("**/*.rej"))
     assert "Project directory may have *.rej files" in result.stdout
-    assert "Retrying again with a different update stratergy." in result.stdout
+    assert "Retrying again with a different update strategy." in result.stdout
 
 
 def test_update_interactive_cancelled(cruft_runner, cookiecutter_dir):